[firebase_remote_config] Support v2 android embedder. (#282)
