+++ packages/firebase_remote_config/example/pubspec.yaml
@@ -10,7 +10,7 @@ dependencies:
   cupertino_icons: ^0.1.0
   firebase_remote_config:
     path: ../
-  firebase_core: ^0.4.0
+  firebase_core: ^0.4.1
 
 dev_dependencies:
   flutter_test:
@@ -18,3 +18,9 @@ dev_dependencies:
   flutter_driver:
     sdk: flutter
   test: any
+  e2e: ^0.2.0
+
+environment:
+  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  flutter: ">=1.9.1+hotfix.2 <2.0.0"
+
