+++ packages/firebase_remote_config/example/test_driver/firebase_remote_config_e2e.dart
@@ -1,12 +1,9 @@
-import 'dart:async';
-import 'package:flutter_driver/driver_extension.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:e2e/e2e.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
 
 void main() {
-  final Completer<String> completer = Completer<String>();
-  enableFlutterDriverExtension(handler: (_) => completer.future);
-  tearDownAll(() => completer.complete(null));
+  E2EWidgetsFlutterBinding.ensureInitialized();
 
   group('$RemoteConfig', () {
     RemoteConfig remoteConfig;
@@ -21,7 +18,7 @@ void main() {
       });
     });
 
-    test('fetch', () async {
+    testWidgets('fetch', (WidgetTester tester) async {
       final DateTime lastFetchTime = remoteConfig.lastFetchTime;
       expect(lastFetchTime.isBefore(DateTime.now()), true);
       await remoteConfig.fetch(expiration: const Duration(seconds: 0));
