--- packages/firebase_remote_config/example/test_driver/firebase_remote_config_test.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_driver/flutter_driver.dart';
-
-Future<void> main() async {
-  final FlutterDriver driver = await FlutterDriver.connect();
-  await driver.requestData(null, timeout: const Duration(minutes: 1));
-  driver.close();
-}
